fix(ARC-560): test crash e2e message
