+++ apps/web/src/shared/hooks/useIsMounted.ts
@@ -1,0 +1,11 @@
+import { useSyncExternalStore } from 'react';
+
+const emptySubscribe = () => () => {};
+
+export function useIsMounted(): boolean {
+  return useSyncExternalStore(
+    emptySubscribe,
+    () => true,
+    () => false,
+  );
+}
