+++ apps/web/src/widgets/header/ui/useIsMounted.ts
@@ -1,11 +1,1 @@
-import { useSyncExternalStore } from 'react';
-
-const emptySubscribe = () => () => {};
-
-export function useIsMounted(): boolean {
-  return useSyncExternalStore(
-    emptySubscribe,
-    () => true,
-    () => false,
-  );
-}
+export { useIsMounted } from '@/shared/hooks/useIsMounted';
